requirements.txt
flask
python-nmap
pandas
scikit-learn
joblib
fpdf
